+++ frontends/web/src/routes/bitsurance/bitsurance.tsx
@@ -3,61 +3,66 @@
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { TAccount } from '@/api/account';
+import { AccountCode, TAccount } from '@/api/account';
 import { bitsuranceLookup } from '@/api/bitsurance';
+import { Header, GuidedContent, GuideWrapper, Main } from '@/components/layout';
+import { MarketTab } from '@/routes/market/components/markettab';
 import { alertUser } from '@/components/alert/Alert';
 import { A } from '@/components/anchor/anchor';
 import { Button } from '@/components/forms';
 import { Checked, Sync, SyncLight } from '@/components/icon';
-import { Column, ColumnButtons, GuidedContent, GuideWrapper, Header, Main, ResponsiveGrid } from '@/components/layout';
+import { Column, ColumnButtons, ResponsiveGrid } from '@/components/layout';
 import { View, ViewContent } from '@/components/view/view';
 import { useDarkmode } from '@/hooks/darkmode';
-import { BitsuranceGuide } from './guide';
 import { i18n } from '@/i18n/i18n';
+import { BitsuranceGuide } from './guide';
 import style from './bitsurance.module.css';
 
 type TProps = {
   accounts: TAccount[];
+  code: AccountCode;
 };
 
-export const Bitsurance = ({ accounts }: TProps) => {
+export const Bitsurance = ({ accounts, code }: TProps) => {
   const navigate = useNavigate();
   const { t } = useTranslation();
   const { isDarkMode } = useDarkmode();
+
   const [insuredAccounts, setInsuredAccounts] = useState<TAccount[]>([]);
-  const [redirecting, setRedirecting] = useState(true);
   const [scanDone, setScanDone] = useState(false);
   const [scanLoading, setScanLoading] = useState(false);
 
   const amount = '100.000€';
+  const hasBitsuranceAccount = accounts.some(({ bitsuranceStatus }) => bitsuranceStatus);
 
   useEffect(() => {
-    if (accounts.some(({ bitsuranceStatus }) => bitsuranceStatus)) {
+    if (code && hasBitsuranceAccount) {
       // replace current history item when redirecting so that the user can go back
-      navigate('/bitsurance/dashboard', { replace: true });
-    } else {
-      setRedirecting(false);
+      navigate(`/market/bitsurance/dashboard/${code}`, { replace: true });
     }
 
     return () => setInsuredAccounts([]);
-  }, [accounts, navigate]);
+  }, [code, hasBitsuranceAccount, navigate]);
 
   const detect = async (redirectToDashboard: boolean) => {
     setScanLoading(true);
     setScanDone(false);
     setInsuredAccounts([]);
-    const response = await bitsuranceLookup();
-    if (!response.success) {
-      alertUser(response.errorMessage);
-      return;
-    }
-    const insuredAccountsCodes = response.bitsuranceAccounts.map(account => account.status ? account.code : null);
-    const insured = accounts.filter(({ code }) => insuredAccountsCodes.includes(code));
-    setInsuredAccounts(insured);
-    setScanDone(true);
-    setScanLoading(false);
-    if (insured.length && redirectToDashboard) {
-      navigate('/bitsurance/dashboard');
+    try {
+      const response = await bitsuranceLookup();
+      if (!response.success) {
+        alertUser(response.errorMessage);
+        return;
+      }
+      const insuredAccountsCodes = response.bitsuranceAccounts.map(account => account.status ? account.code : null);
+      const insured = accounts.filter(({ code }) => insuredAccountsCodes.includes(code));
+      setInsuredAccounts(insured);
+      setScanDone(true);
+      if (insured.length && redirectToDashboard) {
+        navigate(`/market/bitsurance/dashboard/${code}`);
+      }
+    } finally {
+      setScanLoading(false);
     }
   };
 
@@ -74,10 +79,10 @@ export const Bitsurance = ({ accounts }: TProps) => {
     // we force a detection to verify if there is any new insured account
     // before proceeding to the next step.
     await detect(false);
-    navigate('/bitsurance/account');
+    navigate(`/market/bitsurance/account/${code}`);
   };
 
-  if (redirecting) {
+  if (hasBitsuranceAccount) {
     return null;
   }
 
@@ -85,7 +90,12 @@ export const Bitsurance = ({ accounts }: TProps) => {
     <Main>
       <GuideWrapper>
         <GuidedContent>
-          <Header title={<h2>{t('sidebar.insurance')}</h2>} />
+          <Header title={<h2>{t('generic.buySell')}</h2>} />
+          <MarketTab
+            accounts={accounts}
+            activeTab="insure"
+            code={code}
+          />
           <View fullscreen={false}>
             <ViewContent>
               <p className={style.noVspace}>{t('bitsurance.intro.text1', { amount })}</p>
@@ -141,7 +151,7 @@ export const Bitsurance = ({ accounts }: TProps) => {
             </ViewContent>
           </View>
         </GuidedContent>
-        <BitsuranceGuide/>
+        <BitsuranceGuide />
       </GuideWrapper>
     </Main>
   );